Create a single subroutine that determines whether a file is a "reserved" file used by Fossil itself, or is potentially a valid repository file. This processing used to be duplicated at each place where it was needed.
FossilOrigin-Name: 56d69dbd858083354713052ea7be24cb0e711e06

+++ src/checkout.c
@@ -263,22 +263,16 @@ void checkout_cmd(void){
 /*
 ** Unlink the local database file
 */
-void unlink_local_database(void){
-  static const char *azFile[] = {
-     "%s_FOSSIL_",
-     "%s_FOSSIL_-journal",
-     "%s_FOSSIL_-wal",
-     "%s_FOSSIL_-shm",
-     "%s.fos",
-     "%s.fos-journal",
-     "%s.fos-wal",
-     "%s.fos-shm",
-  };
+static void unlink_local_database(int manifestOnly){
+  const char *zReserved;
   int i;
-  for(i=0; i<sizeof(azFile)/sizeof(azFile[0]); i++){
-    char *z = mprintf(azFile[i], g.zLocalRoot);
-    unlink(z);
-    free(z);
+  for(i=0; (zReserved = fossil_reserved_name(i))!=0; i++){
+    if( manifestOnly==0 || zReserved[0]=='m' ){
+      char *z;
+      z = mprintf("%s%s", g.zLocalRoot, zReserved);
+      unlink(z);
+      free(z);
+    }
   }
 }
 
@@ -297,6 +291,7 @@ void close_cmd(void){
   if( !forceFlag && unsaved_changes()==1 ){
     fossil_fatal("there are unsaved changes in the current checkout");
   }
+  unlink_local_database(1);
   db_close();
-  unlink_local_database();
+  unlink_local_database(0);
 }
